Change structure to folders for dev and app, save model's state dict
